% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vietnamcode_data.R
\docType{data}
\name{vietnamcode_data}
\alias{vietnamcode_data}
\title{Vietnam provinces ID}
\format{A data frame with X rows and X variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
vietnamcode_data
}
\description{
Vietnam provinces ID
}
\keyword{datasets}

