% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vietnamcode_data.R
\docType{data}
\name{vietnamcode_data}
\alias{vietnamcode_data}
\title{Vietnam provinces ID}
\format{A data frame

\itemize{
  \item{province_name}{province name}
  \item{province_name}{province name with diacritics}
  \item{pci}{used by the Provincial Competitiveness Index}
  \item{enterprise_census}{used by the General Statistics Office (GSO) for the Enterprise Census}
  \item{enterprise_census_old}{old GSO code}
  \item{enterprise_census_c}{GSO code as character}
  \item{regex}{Regular Expression to match different spellings}
}}
\usage{
vietnamcode_data
}
\description{
Vietnam provinces ID
}
\keyword{datasets}

